Implemented the basic structure of the STNU.
In this file you can find two structures I implemented for representing
the edges, and various representations of the graph (which are for now,
subject to change).

There are two methods that will  be useful for reading data, one of them
I will implement later.

@abi: We need a map from string -> int for the labels, I left a TODO in
the code for you.

@david: I left a comment where I describe how to iterate over the edges
for bellman.